Fix more defects detected by clang at high warnings level
git-svn-id: svn://svn.code.sf.net/p/sobjectizer/repo/branches/so_5/5.5.13--dev-trunk--no-rt-ns@1998 1d1731ae-3f85-447d-88f8-c72b288064d1

+++ dev/sample/so_5/hardwork_imit/main.cpp
@@ -252,9 +252,9 @@ parse_params( int argc, char ** argv )
 		};
 
 	if( 2 < argc )
-		r.m_requests = std::atoi( argv[ 2 ] );
+		r.m_requests = static_cast< unsigned int >( std::atoi( argv[ 2 ] ) );
 	if( 3 < argc )
-		r.m_milliseconds = std::atoi( argv[ 3 ] );
+		r.m_milliseconds = static_cast< unsigned int >( std::atoi( argv[ 3 ] ) );
 
 	std::cout << "Config:\n"
 		"\t" "dispatcher: " << argv[ 1 ] << "\n"
